0.9.0 commit

+++ Components/OptionScreen.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {View, StyleSheet, Text, AsyncStorage} from 'react-native';
+import {View, StyleSheet, Text, AsyncStorage, Alert} from 'react-native';
 import {Card, CardItem, Switch, Picker} from 'native-base';
 
 import ChoiceScreen from './ChoiceScreen';
@@ -11,6 +11,7 @@ class OptionScreen extends React.Component {
 
     this.state = {
       isDownload: false,
+      isSave: false,
       mapSearchLimit: '100',
       db: new GGDB(),
     };
@@ -19,12 +20,13 @@ class OptionScreen extends React.Component {
   }
 
   initStatus = async () => {
-    let isDownload = await AsyncStorage.getItem('isDownload');
     let mapSearchLimit = await AsyncStorage.getItem('mapSearchLimit');
-    // console.log(isDownload);
-    // console.log(isDownload === 'true');
+    let isDownload = await AsyncStorage.getItem('isDownload');
+    let isSave = await AsyncStorage.getItem('isSave');
+
     this.setState({
       isDownload: isDownload === 'true',
+      isSave: isSave === 'true',
       mapSearchLimit: mapSearchLimit,
     });
   };
@@ -32,13 +34,66 @@ class OptionScreen extends React.Component {
   downloadToggle = async () => {
     const {db, isDownload} = this.state;
     if (isDownload) {
-      await AsyncStorage.removeItem('isDownload');
-      await db.deleteGgmoneyAll();
-      await db.initMstSigun();
+      Alert.alert(
+        '다운로드 항목 제거 확인',
+        '다운로드받은 데이터도 제거하시겠습니까?',
+        [
+          {
+            text: '예',
+            onPress: async () => {
+              await db.deleteGgmoneyAll();
+              await db.initMstSigun();
+              await AsyncStorage.removeItem('isDownload');
+              await AsyncStorage.removeItem('isSave');
+              this.setState({isDownload: false, isSave: false});
+            },
+          },
+          {
+            text: '아니오',
+            onPress: async () => {
+              await AsyncStorage.removeItem('isDownload');
+              await AsyncStorage.removeItem('isSave');
+              this.setState({isDownload: false, isSave: false});
+            },
+          },
+          {
+            text: '취소',
+            // onPress: () => {
+            //   console.log('Cancel Pressed');
+            // },
+            style: 'cancel',
+          },
+        ],
+        {cancelable: false},
+      );
     } else {
+      Alert.alert(
+        '데이터 다운로드',
+        '표시되는 리스트에서 시군을 선택해주세요.',
+        [
+          {
+            text: '확인',
+            style: 'default',
+          },
+        ],
+        {cancelable: false},
+      );
+
       AsyncStorage.setItem('isDownload', 'true');
+      AsyncStorage.setItem('isSave', 'true');
+      this.setState({isDownload: true, isSave: true});
     }
-    this.setState({isDownload: !isDownload});
+  };
+
+  saveToggle = async () => {
+    const {isSave} = this.state;
+
+    if (isSave) {
+      await AsyncStorage.removeItem('isSave');
+    } else {
+      AsyncStorage.setItem('isSave', 'true');
+    }
+    this.setState({isSave: !isSave});
   };
 
   onMapSearchLimitChange = value => {
@@ -47,14 +102,14 @@ class OptionScreen extends React.Component {
   };
 
   render() {
-    const {isDownload, mapSearchLimit} = this.state;
+    const {isDownload, mapSearchLimit, isSave} = this.state;
     return (
       <View style={styles.container}>
         <View style={styles.topContainer}>
           <Card>
             <CardItem>
               <View>
-                <View style={{flexDirection: 'row'}}>
+                <View style={{flexDirection: 'row', alignItems: 'center'}}>
                   <Text style={styles.cardItemText}>지도 검색시 개수 제한</Text>
 
                   <Picker
@@ -69,13 +124,22 @@ class OptionScreen extends React.Component {
                     <Picker.Item label="1000 개" value="1000" />
                   </Picker>
                 </View>
+                <View>
+                  <Text style={styles.cardItemSubText}>
+                    지도에 너무 많은 마커가 표시되지 않도록 제한하는 기능입니다.
+                  </Text>
+                  <Text style={styles.cardItemSubText}>
+                    기기의 성능을 고려하여 적절한 수로 설정해주세요.
+                  </Text>
+                  <Text style={styles.cardItemSubText}>(Default : 100개)</Text>
+                </View>
               </View>
             </CardItem>
           </Card>
           <Card>
             <CardItem>
               <View>
-                <View style={{flexDirection: 'row'}}>
+                <View style={{flexDirection: 'row', alignContent: 'center'}}>
                   <Text style={styles.cardItemText}>다운로드 방식 사용</Text>
                   <Switch
                     onValueChange={this.downloadToggle}
@@ -92,6 +156,26 @@ class OptionScreen extends React.Component {
                   </Text>
                   <Text style={styles.cardItemSubText}>
                     키워드 검색으로 빠르고 편리한 검색이 가능합니다.
+                  </Text>
+                </View>
+              </View>
+            </CardItem>
+            <CardItem>
+              <View>
+                <View style={{flexDirection: 'row', alignContent: 'center'}}>
+                  <Text style={styles.cardItemText}>마지막 검색어 저장</Text>
+                  <Switch
+                    onValueChange={this.saveToggle}
+                    value={isSave}
+                    disabled={!isDownload}
+                  />
+                </View>
+                <View>
+                  <Text style={styles.cardItemSubText}>
+                    마지막으로 검색한 검색어를 저장합니다.
+                  </Text>
+                  <Text style={styles.cardItemSubText}>
+                    다운로드 방식 사용시 활성화됩니다.
                   </Text>
                 </View>
               </View>
